create BaseFavoritingToolBannerViewModel
